+++ module-api/src/main/java/com/welcommu/moduleapi/user/UserController.java
@@ -4,10 +4,11 @@ import com.welcommu.modulecommon.dto.ApiResponse;
 import com.welcommu.moduledomain.auth.AuthUserDetailsImpl;
 import com.welcommu.moduledomain.company.CompanyRole;
 import com.welcommu.moduledomain.user.User;
+import com.welcommu.moduleservice.user.UserService;
 import com.welcommu.moduleservice.user.dto.UserModifyRequest;
 import com.welcommu.moduleservice.user.dto.UserRequest;
 import com.welcommu.moduleservice.user.dto.UserResponse;
-import com.welcommu.moduleservice.user.UserService;
+import com.welcommu.moduleservice.user.UserServiceImpl;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.AllArgsConstructor;
